Skip event 0 if it has valid data
Event 0 is a reserved id meaning `This Event` and it should never be used directly,
it is possible to have an event in this slot and have it work, however due to hard
coded behavior it is ill-advised to use this slot because it can cause unintended
side-effects that are hard to debug, the extent of which are unknown.

Ignoring id 0 matches MV/MZ behavior and is technically correct even if the reasoning
is rather contrived.