Add map status to Database' isModified
I've temporarily wrapped the code in CHECK_EXPERIMENTAL to prevent performance issues,
my next major project is to get all of the isModified signals wired up and into Database
so that a single change coalesces into the main application.

That approach will allows us to be *told* the project has been modified rather than being
a pushy date on prom-night.